Make small fixes
