+++ app/src/main/java/com/simply/birthdayapp/presentation/navigation/RootNavigation.kt
@@ -5,6 +5,7 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import com.simply.birthdayapp.presentation.ui.screens.auth.AuthMainScreen
+import com.simply.birthdayapp.presentation.ui.screens.main.MainScreen
 
 @Composable
 fun RootNavigation() {
@@ -14,6 +15,9 @@ fun RootNavigation() {
         navController = navController,
         startDestination = RootDestination.AuthMainScreen.route,
     ) {
+        composable(RootDestination.MainScreen.route){
+            MainScreen()
+        }
         composable(RootDestination.AuthMainScreen.route) {
             AuthMainScreen()
         }
@@ -21,5 +25,5 @@ fun RootNavigation() {
 }
 sealed class RootDestination(val route: String) {
     data object MainScreen : RootDestination("main-screen")
-    object AuthMainScreen : RootDestination("landingScreen")
+    data object AuthMainScreen : RootDestination("landing-Screen")
 }
