+++ app/src/main/java/com/simply/birthdayapp/presentation/ui/screens/auth/register/RegisterScreen.kt
@@ -30,6 +30,6 @@ fun RegisterScreen() {
 
 @Preview(showBackground = false)
 @Composable
-fun RegisterScreenPreview() {
+private fun RegisterScreenPreview() {
     RegisterScreen()
 }
