+++ app/src/main/java/com/simply/birthdayapp/presentation/ui/screens/auth/signIn/SingnInScreen.kt
@@ -28,9 +28,6 @@ fun SignInScreen() {
 
 @Preview(showBackground = true)
 @Composable
-fun RegisterScreenPreview() {
+private fun RegisterScreenPreview() {
     SignInScreen()
 }
-
-
-
